src: Wrap internal data in Arc<Mutex<_>>
They might be wrapped too much sometimes, but it makes it cleaner to
pass them along and make sure they can be used from multiple threads.
Also make some of the functions return copy of the data, just to make it
easier for later use.  Few extra copies do not hurt as much.

This makes it almost pointless to sort the returned warnings in tests/,
but this way it will be more future proof in case that sorting gets
removed/changed later.

Signed-off-by: Martin Kletzander <nert.pinx@gmail.com>