validators_python: Remove vl reference from globals
Since the PyDict is Bound<'py> it keeps the reference essentially
forever (or until the python is finalized, which we do not want to do)
and because that variable has a properly reference counted libvirt
connection the reference counting is off once the VirtLint object is
destroyed.

This fixes the reference counting issue to make tests, which also run
python validators with connections, work properly.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>